rename libalgebra
